refactor: Keep interactions/ to interactions only, behind a barrel
interactions/ now holds nothing but the interactFor functions,
re-exported through index.ts, so consumers import from
lib/interactions/index.js without knowing the file-per-interaction
layout. The prompt framework powering them — PromptClient, the
terminal and memory clients, and the set/reset slot — moves out to sit
beside the directory as lib/prompt.ts and lib/memory-prompt.ts.
