+++ src/lib/commands/local/login.ts
@@ -2,7 +2,7 @@ import { assertMutable, login } from 'lib/auth/operations.js'
 import type { Command } from 'lib/commands/registry.js'
 import { stringFlag } from 'lib/commands/spec.js'
 import { NonInteractiveError } from 'lib/errors.js'
-import { interactForLogin } from 'lib/interactions/login.js'
+import { interactForLogin } from 'lib/interactions/index.js'
 
 export const loginCommand: Command = {
   definition: {
