+++ src/lib/interactions/command-selection.ts
@@ -2,7 +2,7 @@ import { isDeepStrictEqual as isEqual } from 'node:util'
 
 import type { Interactivity } from 'lib/args/parse.js'
 import { NonInteractiveError, PromptCancelledError } from 'lib/errors.js'
-import { promptAutocomplete, withBackHint } from './prompt.js'
+import { promptAutocomplete, withBackHint } from 'lib/prompt.js'
 
 const uniqBy = <T>(items: T[], keyOf: (item: T) => unknown): T[] => {
   const seen = new Set<unknown>()
